Moving CandidateTestRun component and CandidateModals under CandidateTests tree
