+++ client/components/CandidateTests/CandidateModals/ThankYouModal/ThankYouModal.css
